sep(heredc): sep open_heredoc_fd() in new file
