+++ srcs/parse/open_heredoc_fd.c
@@ -1,0 +1,40 @@
+#include <fcntl.h>
+#include <readline/readline.h>
+#include "ms_parse.h"
+#include "ft_dprintf.h"
+#include "ft_sys.h"
+#include "ft_mem.h"
+
+static void	clean_up_prev_heredoc(int in_fd, char *filename)
+{
+	close(in_fd); 		// todo: error handling
+	unlink(filename);	// todo: error handling
+	ft_free(&filename);
+}
+
+static int	open_file_dup_errno(const char *file, int *tmp_err)
+{
+	int	fd;
+
+	errno = 0;
+	fd = open(file, O_CREAT | O_WRONLY | O_TRUNC, 0664);
+//	ft_dprintf(2, "new in_fd:%d, filename:%s\n", *in_fd, *filename);
+	*tmp_err = errno;
+	return (fd);
+}
+
+t_result	open_heredoc_filedes(int *in_fd, char **filename)
+{
+	int	tmp_err;
+
+	if (*in_fd != IN_FD_INIT)
+		clean_up_prev_heredoc(*in_fd, *filename);
+	*filename = create_heredoc_filename();
+	*in_fd = open_file_dup_errno(*filename, &tmp_err);
+	if (tmp_err == ENOMEM)
+	{
+		perror("open"); // todo: msg
+		return (PROCESS_ERROR);
+	}
+	return (SUCCESS);
+}
